Rename sdi-session-helpers to sdi-user-session-manager
