+++ src/sdi-user-session-manager.c
@@ -15,7 +15,7 @@
  *
  */
 
-#include "sdi-session-helpers.h"
+#include "sdi-user-session-manager.h"
 
 static Login1Manager *login_manager = NULL;
 static guint idle_id = 0;
