+++ src/sdi-user-session-manager.h
